索引位置超过int的最大限制导致数据定位错误问题修改

+++ src/main/scala/org/apache/spark/sql/gdb/core/IndexInfo.scala
@@ -2,6 +2,6 @@ package org.apache.spark.sql.gdb.core
 
 /**
   */
-case class IndexInfo(var objectID: Int, var seek: Int) {
+case class IndexInfo(var objectID: Int, var seek: Long) {
   def isSeekable = seek > 0
 }
